+++ src/ui/app.react.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import Sidebar from './sidebar'
 import Header from './header.react'
+import Footer from './footer/index.react'
 import SendForm from './send-form'
 
 let App = React.createClass({
@@ -10,11 +11,11 @@ let App = React.createClass({
     return (
       <div id='container' >
         <Sidebar />
-
         <section id='main-content'>
           <Header />
           <SendForm />
         </section>
+        <Footer />
       </div>
     )
   }